% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca_moran.R
\name{cca_moran}
\alias{cca_moran}
\title{Moran's I Test for Spatial Dependence in Canonical Variates}
\usage{
cca_moran(U1, V1, loc_id, lw)
}
\arguments{
\item{U1}{numeric vector, the first canonical variate from X}

\item{V1}{numeric vector, the first canonical variate from Y}

\item{loc_id}{vector, location ID for each observation}

\item{lw}{listw object, spatial weights}
}
\value{
list of Moran's I test results for U1 and V1
}
\description{
Moran's I Test for Spatial Dependence in Canonical Variates
}
