% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canonical_variates_local.R
\name{cca_variates_local}
\alias{cca_variates_local}
\title{Compute the Influence of a Certain Canonical Variate in Different Locations}
\usage{
cca_variates_local(U1, V1, loc_id, lw_mat)
}
\arguments{
\item{U1}{numeric vector, the first canonical variate from X}

\item{V1}{numeric vector, the first canonical variate from Y}

\item{loc_id}{vector, location ID for each observation}

\item{lw_mat}{matrix, spatial weights matrix}
}
\value{
list with local U1 and V1 variates
}
\description{
Compute the Influence of a Certain Canonical Variate in Different Locations
}
