% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho_local.R
\name{rho_local}
\alias{rho_local}
\title{Compute the Influence of Neighbors Towards the Global CCA Coefficient}
\usage{
rho_local(global_cor, lw_mat)
}
\arguments{
\item{global_cor}{numeric, the i-th canonical correlation from global (regular) CCA}

\item{lw_mat}{matrix, spatial weights matrix}
}
\value{
matrix of local canonical correlations
}
\description{
Compute the Influence of Neighbors Towards the Global CCA Coefficient
}
